requirements.txt
annotated-types==0.6.0
anyio==3.7.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.4
cryptography==41.0.5
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
ecdsa==0.18.0
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.104.1
fonttools==4.28.2
h11==0.14.0
html5lib==1.1
idna==3.4
ipykernel==6.5.1
ipython==7.30.0
jedi==0.18.1
jupyter-client==7.1.0
jupyter-core==4.9.1
kiwisolver==1.3.2
lxml==4.9.2
matplotlib==3.5.0
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
numpy==1.21.4
openpyxl==3.0.9
packaging==21.3
pandas==1.3.4
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.23
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.6
pytz==2021.3
pywin32==302
pyzmq==22.3.0
requests==2.31.0
rsa==4.9
scipy==1.7.3
seaborn==0.11.2
setuptools-scm==6.3.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
tomli==1.2.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.8.0
urllib3==2.0.3
uvicorn==0.24.0.post1
wcwidth==0.2.5
webencodings==0.5.1
